refac: abstract team porfolios comp
