+++ libs/features/src/lib/project/components/organisms/index.ts
@@ -1,3 +1,4 @@
 export * from './project-card/project-card';
 export * from './company-portfolios/company-portfolios';
 export * from './personal-portfolios/personal-portfolios';
+export * from './team-portfolios/team-portfolios';
