+++ libs/features/src/lib/project/components/organisms/team-portfolios/team-portfolios.spec.tsx
@@ -1,0 +1,12 @@
+import { render } from '@testing-library/react';
+
+import TeamPortfolios from './team-portfolios';
+
+describe('TeamPortfolios', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(
+      <TeamPortfolios />
+    );
+    expect(baseElement).toBeTruthy();
+  });
+});
